🏗️ (refactor): Move previous framework details state to a view model
Created a view model for the framework details screen - FrameworkDetailViewModel and moved the previous framework and safari state properties to this file. In both the framework detail grid & list views, the view model is created as an @ObservedObject as it still requires the framework data from their previous views, such as when the user tapped on a selected framework to view that framework details screen.