♻️ (refactor): Update Learn More button to a Link
Refactored the previous Learn More button in the framework detail view to use the built-in SwiftUI Link that was made available from iOS 14+. Some behaviour changes to note: the new Link now opens the Safari app and directs the user to the selected framework URL home page. This was not the case with the previous UIKit integration of SafariViewController, where tapping on the button instead opens an in-app Safari window to the framework URL home page. More details on this refactor decision will be made clear with a new dedicated section in the project README documentation.