⚰️ (chore): Remove UIKit integration of SafariViewController
The previous UIKit integration of SafariViewController source code was removed due to its redundancy, as the previously refactored button in the Framework detail views now uses the built-in SwiftUI Link. The main behaviour change that affects to UX of the app is the initialization of the Safari app itself to open the URL link to the framework homepage, compare to the previous behaviour where an in-app Safari window simply opens and direct the user to the selected framework URL link homepage. More details will be included for this refactor decision in the project README documentation.