🚸 (new): Change previous sample framework to a "not found" framework object
Changed the sample framework details in previews and optional unwrappings to display a "not-found" framework object with the associated details. This change makes for a better UX in the event if the framework data is missing, the framework detail screen will display the appropriate details informing the userof what happened instead of simply showing a sample framework with no relevant context.